feat!: no longer inline telemetry file into zip as telemetry is now handled directly within the updated serverless-functions-api package
The telemetry implementation within serverless-functions-api is no longer a separate published file that we need to import. The implementation is now contained within the library and exposed to consumers via the new `@netlify/otel` package.

The tracing implementation is off by default, and can be turned on for a single request by setting the header 'x-nf-enable-tracing'

We are including the telemetry implementation within serverless-functions-api for a few reason:
- This approach resolves a bug where a customer's own use of the otel libraries would interact with our use of otel due to otel being implemented as a singleton. The new version of serverless-functions-api has bundled otel, which will stop these scenario from being possible.
- We can now activate otel on a per request basis, which helps our support team as they can now gather otel data without asking the customer to redeploy their site
- We are populating the root context with a Span ID which will be supplied by Proxy - this will enable Netlify to connect our otel from ingress (Stargate) all the way through our infrastructure and to the customers application code.